Moved settings-tab into separate file
